+++ src/geometor/explorer/__init__.py
@@ -10,5 +10,5 @@ GEOMETOR
 __author__ = "GEOMETOR"
 __maintainer__ = "GEOMETOR"
 __email__ = "github@geometor.com"
-__version__ = "0.3.17"
+__version__ = "0.4.0"
 __licence__ = "MIT"